refactor: extract PersistedEditor component and fix serialization
Extract load/save logic into a shared PersistedEditor component that
takes a note id prop. Fix two serialization bugs: use blocks.blocks
getter (not the Blocks instance directly, which has private fields)
and guard against corrupted data on load.
